PSM libs: p9. Fix unittests signed vs unsigned comparison warning
shell-encryption and private_membership leaked some configs
which disable some default warnings.

In order to re-enable all these warnings for chrome/browser/*
in this CL crrev.com/c/3303510, all warnings need to be fixed.

This CL is fixing the comparison between two variables to have
the same type (i.e. either both are signed or unsigned integers).

Bug: chromium:1212837, b:201763017
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3478540
Commit-Queue: Amr Aboelkher <amraboelkher@chromium.org>
Cr-Commit-Position: refs/heads/main@{#976247}

+++ chrome/browser/extensions/external_provider_impl_chromeos_unittest.cc
@@ -124,7 +124,7 @@ class ExternalProviderImplChromeOSTest : public ExtensionServiceTestBase {
     }
   }
 
-  void ValidateExternalProviderCountInAppMode(int expected_count) {
+  void ValidateExternalProviderCountInAppMode(size_t expected_count) {
     base::CommandLine* command = base::CommandLine::ForCurrentProcess();
     command->AppendSwitchASCII(switches::kForceAppMode, std::string());
     command->AppendSwitchASCII(switches::kAppId, std::string("app_id"));
@@ -329,7 +329,7 @@ TEST_F(ExternalProviderImplChromeOSTest, ChromeAppKiosk) {
   fake_user_manager()->AddKioskAppUser(kiosk_account_id);
   fake_user_manager()->LoginUser(kiosk_account_id);
 
-  ValidateExternalProviderCountInAppMode(3);
+  ValidateExternalProviderCountInAppMode(3u);
 }
 
 // Validate the external providers enabled in the ARC++ App Kiosk session. The
@@ -340,7 +340,7 @@ TEST_F(ExternalProviderImplChromeOSTest, ArcAppKiosk) {
   fake_user_manager()->AddArcKioskAppUser(kiosk_account_id);
   fake_user_manager()->LoginUser(kiosk_account_id);
 
-  ValidateExternalProviderCountInAppMode(1);
+  ValidateExternalProviderCountInAppMode(1u);
 }
 
 // Validate the external providers enabled in the Web App Kiosk session. The
@@ -351,7 +351,7 @@ TEST_F(ExternalProviderImplChromeOSTest, WebAppKiosk) {
   fake_user_manager()->AddWebKioskAppUser(kiosk_account_id);
   fake_user_manager()->LoginUser(kiosk_account_id);
 
-  ValidateExternalProviderCountInAppMode(1);
+  ValidateExternalProviderCountInAppMode(1u);
 }
 
 }  // namespace extensions
